♻️ GS13TP task: decompose context ingest prompt
