+++ .agentplane/tasks/202605290600-GS13TP/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 GS13TP task: Context ingest task prompt decomposition [202605290600-GS13TP]
