+++ .agentplane/tasks/202605290600-GS13TP/pr/review.md
@@ -1,0 +1,38 @@
+# PR Review
+
+Created: 2026-05-29T06:00:47.010Z
+
+## Task
+
+- Task: `202605290600-GS13TP`
+- Title: Context ingest task prompt decomposition
+- Status: DOING
+- Branch: `task/202605290600-GS13TP/context-ingest-task-prompt-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290600-GS13TP/README.md`
+
+## Verification
+
+- State: ok
+- Note: Context ingest prompt construction extracted into packages/agentplane/src/context/ingest-task-prompt.ts; createTaskNewParsed behavior and prompt payload are preserved. Checks passed: bun test packages/agentplane/src/commands/context/release-readiness.test.ts (21 pass); bun run typecheck; bun run arch:check; bun run knip:check; bun run lint:core; bun run format:changed; bun run hotspots:check (runtime hotspots 5 -> 4).
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T06:00:47.010Z
+- Branch: task/202605290600-GS13TP/context-ingest-task-prompt-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+ .../agentplane/src/context/ingest-task-prompt.ts   | 142 +++++++++++++++++++++
+ packages/agentplane/src/context/ingest-task.ts     | 139 +-------------------
+ 2 files changed, 146 insertions(+), 135 deletions(-)
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
